% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{collate_paginated_output}
\alias{collate_paginated_output}
\title{Collate paginated output for a given query}
\usage{
collate_paginated_output(
  endpoint,
  query_params,
  pagination_size,
  response_parser
)
}
\arguments{
\item{endpoint}{Endpoint of the API}

\item{query_params}{List of parameters to pass to the API, values must be strings}

\item{pagination_size}{The size of the response returned (the page size). Should be equal to the value in the query_params}

\item{response_parser}{A callable that parses the response into something that can be concatenated}
}
\description{
Collates output for a given query, where the query must be called multiple times
due to pagination in the API. Assumes that there is a parser for the JSON response
from the API.
}
\examples{
\dontrun{
# Construct a query
endpoint <- "measurements"
query_params <- list(
  location = "Melbourne",
  date_from = "2024-05-24",
  date_to = "2024-05-25",
  limit = "100"
)

# Paginate over results
collate_paginated_output(
  endpoint = endpoint,
  query_params = query_params,
  pagination_size = 100,
  response_parser = get_measures
)
}

}
