% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{get_measurements_for_location}
\alias{get_measurements_for_location}
\title{Get air quality measurements for a location}
\usage{
get_measurements_for_location(
  country = NULL,
  city = NULL,
  location = NULL,
  date_from = Sys.Date() - 365,
  date_to = Sys.Date(),
  is_mobile = FALSE,
  max_observations = 1000,
  ...
)
}
\arguments{
\item{country}{Two-character ISO country code}

\item{city}{String for the city of interest}

\item{location}{smaller-geographical identifier (i..e suburb)}

\item{date_from}{Date observations to start (will be converted to UTC)}

\item{date_to}{Date observations will end (will be converted to UTC)}

\item{is_mobile}{Include mobile station types}

\item{max_observations}{Number of observations to be returned per page.
Larger numbers mean less API calls but more single server load.}

\item{...}{Extra API parameters. Note all must be name-value pairs, and all
values must be of type character.}
}
\description{
Gets the set of air quality measurements from the OpenAQ API. You can filter
by location, date, and type of station
}
