% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{set_openaq_api_key}
\alias{set_openaq_api_key}
\title{Set OpenAQ API Key}
\usage{
set_openaq_api_key(api_key)
}
\arguments{
\item{api_key}{API key to set}
}
\description{
Sets your OpenAQ API key for use by the package \code{airpurifyr}
Register your API key at \url{https://explore.openaq.org/register}, and access your
account and get your API key from the dashboard \url{https://explore.openaq.org/account}
}
